arc_impl.h: NUMA node numbers are just ints
Anything else is asking for trouble

+++ include/sys/arc_impl.h
@@ -155,7 +155,7 @@ typedef struct l1arc_buf_hdr {
 	uint8_t			b_byteswap;
 
 	/* What node the ARC header/buffer was allocated on */
-	uint16_t		b_node;
+	int			b_node;
 
 	/* protected by arc state mutex */
 	arc_state_t		*b_state;
